Revert ""PROJECT AND TASK WORKING""
