--- DAY_1/004_EMP_ID (copy).java
@@ -1,9 +1,0 @@
-public class EMP_ID {
-    public static void main(String[] args){
-        int[] EMP_ID =  new int[25];
-        for(int i=0;i<25;i++){
-            EMP_ID[i] = i;
-            System.out.println(EMP_ID[i]);
-        }
-    }
-}
