--- DAY_1/Classes/src/LinkedList.java
@@ -1,5 +1,0 @@
-public class LinkedList {
-public static void main(String[] args){
-    System.out.println("Hello World");
-}
-}
