--- DAY_1/MEMCACHED/src/main/java/org/example/App.java
@@ -1,13 +1,0 @@
-package org.example;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
