--- DAY_1/PMS/resources/env.sh
@@ -1,7 +1,0 @@
-#! /bin/bash
-export DB_NAME="TEST"  
-export DB_HOST="localhost"
-export UNAME="kouth"
-export PASS="kouth1234"
-export REDIS_HOST="localhost"
-export REDIS_PORT="6379"
