--- DAY_1/classes.java
@@ -1,12 +1,0 @@
-public class classes {
-    int i;
-    String Name;
-    classes(int i,String s){
-        i = i;
-        Name = s;
-    }
-}
-
-public class j{
-    classes obj = new classes();
-}
