--- DAY_1/untitled/src/Lists.java
@@ -1,14 +1,0 @@
-import java.util.*;
-public class Lists {
-public static void main(String args[]){
-        List linkedlist = new LinkedList();
-        List arrayList  = new ArrayList();
-        List vec_list   = new Vector();
-        List stck_list  = new Stack();
-
-
-
-
-
-    }
-}
