--- DAY_1/xml/src/CollToXML.java
@@ -1,24 +1,0 @@
-//import org.jdom2.Document;
-//import org.jdom2.Element;
-//import org.jdom2.JDOMException;
-//import org.jdom2.input.SAXBuilder;
-//
-//import java.io.File;
-//import java.io.FileWriter;
-//import java.io.IOException;
-//import java.io.OutputStream;
-//import java.util.List;
-//import java.util.ArrayList;
-//import src.XMLOBJ;
-//public class CollToXML {
-//    private static final String FILENAME = "src/staff.xml";
-//public static void main(String args[])throws IOException{
-//    ArrayList<XMLOBJ> arr = new ArrayList<XMLOBJ>();
-//    FileWriter f = new FileWriter(FILENAME);
-//    CollXml(arr,System.out,f);
-//}
-//    private static void CollXml(ArrayList<XMLOBJ> arr ,OutputStream output, FileWriter f) throws IOException {
-//        Document doc = new Document();
-//
-//    }
-//}
